Split `crypto.rs` into modular files
