+++ src/crypto/mod.rs
@@ -1,1 +1,3 @@
-pub mod crypto;
+pub mod okamoto_uchiyama;
+pub mod private_key;
+pub mod public_key;
